requirements.txt
aniso8601==9.0.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
flask-restx==1.3.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
PyJWT==2.8.0
pymongo==4.7.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.0
rpds-py==0.18.0
Werkzeug==3.0.2
